cstack/requirements.txt
django==4.0
psycopg2-binary==2.9.9
Pillow==10.3.0